make convoscope main page work as fragment
