+++ android_app/app/src/main/java/com/teamopensmartglasses/convoscope/Config.java
@@ -4,8 +4,8 @@ public class Config {
     //public static final String serverUrl = "https://1q93wu6qkd.execute-api.us-east-2.amazonaws.com"; //jeremy
     //public static final String serverUrl = "https://9259wtmk89.execute-api.us-east-2.amazonaws.com"; //cayden
     //public static final String serverUrl = "https://m0q3aubk6c.execute-api.us-east-1.amazonaws.com"; //cayden, MIT Media Lab Fluid Interfaces Group
-    //public static final String serverUrl = "https://vpmkebx0cl.execute-api.us-east-2.amazonaws.com/api"; //TOSG BOX
-    public static final String serverUrl = "http://192.168.7.117:8080";
+    public static final String serverUrl = "https://vpmkebx0cl.execute-api.us-east-2.amazonaws.com/api"; //TOSG BOX
+//    public static final String serverUrl = "http://192.168.7.117:8080";
 
     public static final Boolean useDevServer = false;
 }
