+++ android_app/app/src/main/res/layout/convoscope_fragment.xml
@@ -16,7 +16,7 @@
       app:layout_constraintWidth_percent="0.95"
       android:layout_height="0dp"
       app:layout_constraintHeight_default="percent"
-      app:layout_constraintHeight_percent="0.6"
+      app:layout_constraintHeight_percent="0.75"
 
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent" />
@@ -30,7 +30,7 @@
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintHeight_default="percent"
 
-      app:layout_constraintHeight_percent="0.4"
+      app:layout_constraintHeight_percent="0.25"
       app:layout_constraintStart_toStartOf="parent"
       app:layout_constraintWidth_default="percent"
       app:layout_constraintWidth_percent="0.6">
@@ -53,41 +53,19 @@
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHeight_default="percent"
 
-      app:layout_constraintHeight_percent="0.4"
+      app:layout_constraintHeight_percent="0.25"
       app:layout_constraintWidth_default="percent"
       app:layout_constraintWidth_percent="0.403">
-
-    <Switch
-        android:id="@+id/toggleSharingSwitch"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:text="Results Sharing"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/settings_button"
         android:layout_width="115dp"
         android:layout_height="63dp"
-        android:layout_marginTop="24dp"
         android:text="Settings"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.45"
+        app:layout_constraintHorizontal_bias="0.509"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/pick_contact_button" />
-
-    <Button
-        android:id="@+id/pick_contact_button"
-        android:layout_width="120dp"
-        android:layout_height="62dp"
-        android:layout_marginTop="24dp"
-        android:text="Select Contact"
-        android:textColor="#FFFFFF"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toggleSharingSwitch" />
+        tools:layout_editor_absoluteY="16dp" />
 
     <!--    <Button-->
 <!--        android:id="@+id/button2"-->
