+++ android_app/app/src/main/res/values/colors.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
   <color name="colorPrimary">#7500B5</color>
+  <color name="red">#FF0000</color>
+  <color name="green">#00FF00</color>
+  <color name="blue">#0000FF</color>
   <color name="colorPrimaryDark">#380057</color>
   <color name="colorAccent">#D81B60</color>
   <color name="colorLightGray">#BBBBBB</color>
