+++ android_app/settings.gradle
@@ -16,5 +16,5 @@ dependencyResolutionManagement {
 rootProject.name = "Convoscope"
 include ':app'
 
-include ':SGMLib'
-project(':SGMLib').projectDir = new File(rootProject.projectDir, '../../SmartGlassesManager/SGM_android/SmartGlassesManager')
+include ':SmartGlassesManager'
+project(':SmartGlassesManager').projectDir = new File(rootProject.projectDir, '../SmartGlassesManager/SGM_android/SmartGlassesManager')
